+++ src/main/java/net/sf/jabref/gui/IconTheme.java
@@ -33,8 +33,10 @@ public class IconTheme {
     public static Font FONT;
     public static Font FONT_16;
 
-    public static final Color DEFAULT_COLOR = new Color(79, 95, 143);
-    public static final Color DEFAULT_DISABLED_COLOR = new Color(200, 200, 200);
+    //public static final Color DEFAULT_COLOR = new Color(79, 95, 143); // JabRef's default color
+    public static final Color DEFAULT_COLOR = new Color(0xBE, 0x05, 0x05); // Christmas edition
+    //public static final Color DEFAULT_DISABLED_COLOR = new Color(200, 200, 200); // JabRef's default color
+    public static final Color DEFAULT_DISABLED_COLOR = new Color(0x00, 0x99, 0x00);
     public static final int DEFAULT_SIZE = 24;
     public static final int SMALL_SIZE = 16;